Don't inhibit GC during weak table operations except rehash
Since GC might occur between machine instructions that manipulate the
table's freelist from Lisp, it is no longer possible to share the same
freelist between Lisp and C. Adding a second freelist managed via atomic
push/pop on ordinary cons cells solves that problem. One freelist is for
REMHASH, the other is for culled entries. PUTHASH can pop from either.

Also:
* Fix fencepost bug in puthash helper: if the inserted index is exactly
  at the high-water-mark, do NOT increase the high-water-mark.

* Fix spurious printing of "non-fatal corruption" message in scav_vector().
  It could occur when a weak hash-table kv-vector is first created and
  has neither any entries nor a backpointer to the containing table.

* Remove WITH-HASH-TABLE-LOCKS macro

* Rename method slots in hash-tables to fn-IMPL

+++ src/runtime/print.c
@@ -815,7 +815,7 @@ static void print_otherptr(lispobj obj)
         NEWLINE_OR_RETURN;
         if (type >= SIMPLE_ARRAY_UNSIGNED_BYTE_2_WIDETAG &&
             type <= SIMPLE_BIT_VECTOR_WIDETAG)
-            printf("length = %ld", fixnum_value(*ptr));
+            printf("length = %ld", (long)fixnum_value(*ptr));
         else
             printf("Unknown header object?");
         break;
